+++ src/components/TableModule/config.ts
